Extracts serverless mode config tests to config-serverless.test.js
