+++ test/unit/config/config-env.test.js
@@ -595,93 +595,4 @@ tap.test('when overriding configuration values via environment variables', (t) =
       }
     )
   })
-
-  t.test('serverless mode', (t) => {
-    t.autoend()
-
-    t.test('should pick up serverless_mode', (t) => {
-      idempotentEnv({
-        NEW_RELIC_SERVERLESS_MODE_ENABLED: true
-      }, (tc) => {
-        t.equal(tc.serverless_mode.enabled, true)
-        t.end()
-      })
-    })
-
-    t.test('should explicitly disable cross_application_tracer in serverless_mode', (t) => {
-      idempotentEnv({
-        NEW_RELIC_SERVERLESS_MODE_ENABLED: true
-      }, (tc) => {
-        t.equal(tc.serverless_mode.enabled, true)
-        t.equal(tc.cross_application_tracer.enabled, false)
-        t.end()
-      })
-    })
-
-    t.test('should enable DT in serverless_mode when account_id has been set', (t) => {
-      idempotentEnv({
-        NEW_RELIC_SERVERLESS_MODE_ENABLED: true,
-        NEW_RELIC_ACCOUNT_ID: '12345'
-      }, (tc) => {
-        t.equal(tc.serverless_mode.enabled, true)
-        t.equal(tc.distributed_tracing.enabled, true)
-        t.end()
-      })
-    })
-
-    t.test('should not enable distributed tracing when account_id has not been set', (t) => {
-      idempotentEnv({
-        NEW_RELIC_SERVERLESS_MODE_ENABLED: true
-      }, (tc) => {
-        t.equal(tc.serverless_mode.enabled, true)
-        t.equal(tc.distributed_tracing.enabled, false)
-        t.end()
-      })
-    })
-
-    t.test('should default primary_application_id to Unknown when not set', (t) => {
-      idempotentEnv({
-        NEW_RELIC_SERVERLESS_MODE_ENABLED: true,
-        NEW_RELIC_ACCOUNT_ID: '12345'
-      }, (tc) => {
-        t.equal(tc.serverless_mode.enabled, true)
-        t.equal(tc.distributed_tracing.enabled, true)
-
-        t.equal(tc.primary_application_id, 'Unknown')
-        t.end()
-      })
-    })
-
-    t.test(
-      'should set serverless_mode from lambda-specific env var if not set by user',
-      (t) => {
-        idempotentEnv({
-          AWS_LAMBDA_FUNCTION_NAME: 'someFunc'
-        }, (tc) => {
-          t.equal(tc.serverless_mode.enabled, true)
-          t.end()
-        })
-      }
-    )
-
-    t.test('should pick app name from AWS_LAMBDA_FUNCTION_NAME', (t) => {
-      idempotentEnv({
-        NEW_RELIC_SERVERLESS_MODE_ENABLED: true,
-        AWS_LAMBDA_FUNCTION_NAME: 'MyLambdaFunc'
-      }, (tc) => {
-        t.ok(tc.app_name)
-        t.deepEqual(tc.applications(), ['MyLambdaFunc'])
-        t.end()
-      })
-    })
-
-    t.test('should default generic app name when no AWS_LAMBDA_FUNCTION_NAME', (t) => {
-      idempotentEnv({NEW_RELIC_SERVERLESS_MODE_ENABLED: true}, (tc) => {
-        t.ok(tc.app_name)
-        t.deepEqual(tc.applications(), ['Serverless Application'])
-
-        t.end()
-      })
-    })
-  })
 })
